requirements.txt
asgiref==3.5.2
Django==4.1.6
django-cors-headers==3.13.0
django-localflavor==3.1
djangorestframework==3.14.0
install==1.3.5
PyJWT==2.6.0
python-stdnum==1.18
pytz==2022.6
sqlparse==0.4.3
